+++ crates/forge/benches/forge_bench/mod.rs
@@ -1,4 +1,2 @@
 pub mod benchmark;
-mod collect_tests;
-mod compile_tests;
 mod declare_deploy_and_interact;
