--- crates/forge/tests/data/file_reading/.snfoundry_cache/.prev_tests_failed
@@ -1,14 +1,0 @@
-file_reading::tests::non_existent
-file_reading::tests::invalid_quotes
-file_reading::tests::non_ascii
-file_reading::tests::not_number_without_quotes
-file_reading::tests::too_large_number
-file_reading::tests::negative_number
-tests::test::non_existent
-tests::test::invalid_quotes
-tests::test::json_non_existent
-tests::test::non_ascii
-tests::test::negative_number
-tests::test::invalid_json
-tests::test::not_number_without_quotes
-tests::test::too_large_number
